refactor: move detail buttons to component/
