--- app/src/main/java/com/cmput301f20t21/bookfriends/callbacks/OnDetailButtonClick.java
@@ -1,7 +1,0 @@
-package com.cmput301f20t21.bookfriends.callbacks;
-
-import android.view.View;
-
-public interface OnDetailButtonClick {
-    void run(View view);
-}
